first stage move Precompiles inside EvmContext. Create InnerEvmContext
